requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
selenium==4.21.0
webdriver-manager==4.0.1 